refact(search): factor out OrDelimiter
